requirements.txt
flet
supabase
reportlab
pypdf
pycryptodome